requirements.txt
discord_webhook==0.14.0
praw==7.6.0
prawcore==2.3.0
progress==1.6
python-dotenv==0.20.0
requests==2.25.1
